dependencies 공통 모듈화, 전체 문서 작업 보완
