+++ fastapi_template/app/api/dependencies.py
@@ -1,57 +1,22 @@
 """
 # File: fastapi_template/app/api/dependencies.py
 # Description: FastAPI 의존성 주입 관리
-# - 인증 의존성
-# - 데이터베이스 세션 의존성
-# - 권한 검증 의존성
+# - 이 파일은 하위 호환성을 위해 유지되며, 실제 구현은 common/dependencies로 이동됨
 """
 
-from fastapi import Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
-from jose import JWTError
-from sqlalchemy.ext.asyncio import AsyncSession
+from app.common.dependencies import (
+    get_current_user,
+    get_current_active_user,
+    get_current_admin_user,
+    oauth2_scheme,
+)
 
-from app.common.auth import verify_token
-from app.common.config import settings
-from app.common.database import get_db
-from app.common.exceptions import AuthenticationError
-from app.services.user_service import UserService
+# API 고유의 의존성은 여기에 추가할 수 있음
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"{settings.API_V1_STR}/auth/login")
-
-
-async def get_current_user(
-    token: str = Depends(oauth2_scheme), db: AsyncSession = Depends(get_db)
-):
-    """현재 인증된 사용자 정보를 가져오는 의존성 함수"""
-    try:
-        payload = verify_token(token)
-        user_id = payload.get("sub")
-        if not user_id:
-            raise AuthenticationError("Invalid token")
-    except JWTError:
-        raise AuthenticationError("Invalid token")
-
-    user = await UserService.get_user(db, user_id=user_id)
-    if not user:
-        raise AuthenticationError("User not found")
-
-    return user
-
-
-async def get_current_active_user(current_user=Depends(get_current_user)):
-    """활성화된 사용자인지 확인하는 의존성 함수"""
-    if not current_user.is_active:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="Inactive user"
-        )
-    return current_user
-
-
-async def get_current_admin_user(current_user=Depends(get_current_user)):
-    """관리자 권한을 가진 사용자인지 확인하는 의존성 함수"""
-    if not current_user.is_admin:
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN, detail="Not enough privileges"
-        )
-    return current_user
+# 하위 호환성을 위해 모든 의존성을 다시 내보냄
+__all__ = [
+    "get_current_user",
+    "get_current_active_user",
+    "get_current_admin_user",
+    "oauth2_scheme",
+]
