--- fastapi_template/docs/common__overview.md
@@ -1,105 +1,0 @@
-# 공통 모듈 개요 및 색인
-
-이 문서는 FastAPI 템플릿의 공통 모듈에 대한 개요와 상세 문서 링크를 제공합니다.
-
-## 목차
-
-- [인증 모듈](#인증-모듈)
-- [설정 모듈](#설정-모듈)
-- [데이터베이스 모듈](#데이터베이스-모듈)
-- [예외 처리 모듈](#예외-처리-모듈)
-- [스키마 모듈](#스키마-모듈)
-- [유틸리티 모듈](#유틸리티-모듈)
-- [캐싱 모듈](#캐싱-모듈)
-- [보안 모듈](#보안-모듈)
-- [미들웨어 모듈](#미들웨어-모듈)
-- [모니터링 모듈](#모니터링-모듈)
-- [데이터 검증 모듈](#데이터-검증-모듈)
-
-## 인증 모듈
-
-**경로**: `app/common/auth/`
-
-JWT 토큰 기반 인증 및 비밀번호 해싱 기능을 제공합니다.
-
-[인증 모듈 상세 문서](common_auth.md)
-
-## 설정 모듈
-
-**경로**: `app/common/config/`
-
-애플리케이션 구성 설정을 관리하는 모듈입니다.
-
-[설정 모듈 상세 문서](common_config.md)
-
-## 데이터베이스 모듈
-
-**경로**: `app/common/database/`
-
-데이터베이스 연결, 세션 관리, ORM 모델 정의를 위한 모듈입니다.
-
-[데이터베이스 상세 문서](common_database.md)
-
-## 예외 처리 모듈
-
-**경로**: `app/common/exceptions/`
-
-사용자 정의 예외 클래스와 전역 예외 처리 기능을 제공합니다.
-
-[예외 처리 모듈 상세 문서](common_exceptions.md)
-
-## 스키마 모듈
-
-**경로**: `app/common/schemas/`
-
-API 응답, 요청 데이터 검증을 위한 공통 스키마를 제공합니다.
-
-[스키마 모듈 상세 문서](common_schemas.md)
-
-## 유틸리티 모듈
-
-**경로**: `app/common/utils/`
-
-다양한 유틸리티 기능을 제공합니다.
-
-[유틸리티 모듈 상세 문서](common_utils.md)
-
-## 캐싱 모듈
-
-**경로**: `app/common/cache/`
-
-Redis 기반 데이터 캐싱 기능을 제공합니다.
-
-[캐싱 모듈 상세 문서](common_cache.md)
-
-## 보안 모듈
-
-**경로**: `app/common/security/`
-
-데이터 암호화, CORS 설정, 속도 제한 등 보안 관련 기능을 제공합니다.
-
-[보안 모듈 상세 문서](common_security.md)
-
-## 미들웨어 모듈
-
-**경로**: `app/common/middleware/`
-
-요청/응답 처리를 위한 미들웨어 컴포넌트를 제공합니다.
-
-[미들웨어 모듈 상세 문서](common_middleware.md)
-
-## 모니터링 모듈
-
-**경로**: `app/common/monitoring/`
-
-애플리케이션 상태 모니터링 및 로깅 기능을 제공합니다.
-
-[모니터링 모듈 상세 문서](common_monitoring.md)
-
-## 데이터 검증 모듈
-
-**경로**: `app/common/validators/`
-
-이메일, 비밀번호, 파일 등 다양한 데이터 검증 유틸리티를 제공합니다.
-
-[데이터 검증 모듈 상세 문서](common_validators.md)
