+++ fastapi_template/docs/common_modules/common_config.md
