+++ fastapi_template/docs/common_modules/database/04-schemas.md
