+++ fastapi_template/docs/testing/common_test_patterns.md
@@ -1,0 +1,142 @@
+# 공통 테스트 패턴 및 구조
+
+이 문서는 테스트 문서 작성의 일관성을 유지하기 위한 공통 패턴과 템플릿을 정의합니다.
+
+## 모듈 테스트 문서 구조
+
+각 모듈 테스트 가이드는 다음 구조를 따라야 합니다:
+
+```markdown
+# [모듈명] 모듈 테스트 가이드
+
+## 개요
+
+`[모듈명]` 모듈은 [모듈의 주요 목적 및 기능에 대한 간략한 설명]
+
+## 테스트 용이성
+
+- **난이도**: [쉬움|중간|어려움|매우 어려움]
+- **이유**:
+  - [난이도 평가 이유 1]
+  - [난이도 평가 이유 2]
+  - ...
+
+## 테스트 대상
+
+주요 테스트 대상은 다음과 같습니다:
+
+1. **[테스트 대상 카테고리 1]**
+   - [세부 기능 1]
+   - [세부 기능 2]
+   - ...
+
+2. **[테스트 대상 카테고리 2]**
+   - [세부 기능 1]
+   - [세부 기능 2]
+   - ...
+   
+3. ...
+
+## 테스트 접근법
+
+[모듈명] 모듈 테스트 시 다음 접근법을 권장합니다:
+
+1. **[접근법 1]**: [설명]
+2. **[접근법 2]**: [설명]
+3. ...
+
+## 테스트 예시
+
+### [예시 카테고리 1]
+
+```python
+# 테스트 코드 예시
+```
+
+### [예시 카테고리 2]
+
+```python
+# 테스트 코드 예시
+```
+
+## 모킹 전략
+
+[모듈명] 모듈 테스트 시 다음과 같은 모킹 전략을 활용할 수 있습니다:
+
+1. **[모킹 전략 1]**: [설명]
+2. **[모킹 전략 2]**: [설명]
+3. ...
+
+## 테스트 커버리지 확인
+
+[모듈명] 모듈의 테스트 커버리지를 확인하세요:
+
+```bash
+pytest --cov=[모듈 경로] [테스트 경로]
+```
+
+## 모범 사례
+
+1. **[모범 사례 1]**: [설명]
+2. **[모범 사례 2]**: [설명]
+3. ...
+
+## 주의사항
+
+1. **[주의사항 1]**: [설명]
+2. **[주의사항 2]**: [설명]
+3. ...
+
+```
+
+## 공통 테스트 접근법
+
+대부분의 모듈에 적용 가능한 일반적인 테스트 접근법:
+
+1. **단위 테스트**: 개별 함수/메서드/클래스를 독립적으로 테스트
+2. **통합 테스트**: 모듈 간 또는 외부 시스템과의 상호작용 테스트
+3. **기능 테스트**: 특정 기능이나 시나리오의 전체 흐름 테스트
+4. **모의 테스트**: 외부 의존성을 모킹하여 테스트
+5. **파라미터화 테스트**: 다양한 입력값으로 같은 테스트 반복
+6. **예외 테스트**: 오류 상황 및 예외 처리 테스트
+
+## 공통 모킹 전략
+
+여러 모듈 테스트에 공통적으로 적용할 수 있는 모킹 전략:
+
+1. **의존성 대체**: 실제 외부 시스템 대신 테스트용 구현체 사용
+2. **함수/메서드 모킹**: `unittest.mock.patch` 또는 `pytest-mock`을 사용
+3. **컨텍스트 관리자**: `with` 문을 사용한 임시 모킹
+4. **픽스처 기반 모킹**: `pytest` 픽스처를 사용한 재사용 가능한 모킹 설정
+5. **인메모리 대체**: 실제 데이터베이스 대신 인메모리 구현체 사용
+
+## 공통 모범 사례
+
+테스트 작성 시 고려해야 할 일반적인 모범 사례:
+
+1. **격리된 테스트**: 각 테스트는 독립적으로 실행 가능해야 함
+2. **사전 조건 명확화**: 테스트에 필요한 초기 상태를 명확히 설정
+3. **검증 단계 분리**: 준비(Arrange) → 실행(Act) → 검증(Assert) 패턴 사용
+4. **복잡성 관리**: 테스트 코드는 간결하고 이해하기 쉽게 작성
+5. **명확한 테스트 이름**: 테스트 함수 이름은 테스트 목적을 명확히 설명
+6. **테스트 중복 최소화**: 공통 설정은 픽스처나 헬퍼 함수로 추출
+7. **코드 커버리지 관리**: 주요 코드 경로와 예외 상황 모두 테스트
+
+## 공통 주의사항
+
+테스트 작성 시 주의해야 할 일반적인 사항:
+
+1. **상태 관리**: 테스트 간 상태 공유 최소화 및 적절한 정리
+2. **외부 의존성**: 외부 시스템 의존 테스트는 별도 분리하고 스킵 옵션 제공
+3. **테스트 속도**: 테스트는 가능한 빠르게 실행되도록 설계
+4. **비결정적 요소**: 날짜/시간, 무작위성 등 비결정적 요소 제어
+5. **환경 변수**: 환경 변수에 의존하는 테스트는 환경 격리 제공
+
+## 테스트 문서 업데이트 지침
+
+테스트 문서 작성 및 업데이트 시 다음 사항을 지켜주세요:
+
+1. 이 공통 템플릿을 기반으로 각 모듈 테스트 가이드 작성
+2. 모듈별 특성에 맞게 섹션 내용 조정 (추가/삭제 가능)
+3. 실제 예시 코드는 현재 구현에 맞게 최신 상태 유지
+4. 새로운 일반적인 패턴이 발견되면 이 문서에 추가
