+++ fastapi_template/docs/testing/module_guides.md
@@ -1,0 +1,128 @@
+# 모듈별 테스트 가이드
+
+이 문서는 FastAPI 템플릿 프로젝트의 각 모듈을 테스트하는 방법에 대한 가이드라인의 개요를 제공합니다.
+
+> **참고**: 각 모듈별 테스트 가이드는 [공통 테스트 패턴](common_test_patterns.md)에 정의된 구조와 형식을 따릅니다. 새로운 모듈 테스트 가이드 작성 시 이 템플릿을 활용하세요.
+>
+> **위치**: 모든 모듈별 테스트 가이드는 **`modules/`** 디렉토리 내에 위치하고 있습니다.
+
+## 모듈별 테스트 구성
+
+프로젝트의 각 모듈은 테스트 용이성에 차이가 있습니다. 아래는 각 모듈의 테스트 개요와 관련 문서 링크입니다.
+
+### Validators 모듈
+
+- **난이도**: 쉬움
+- **특징**: 순수 함수, 외부 의존성 없음, 결정적 동작
+- **주요 테스트 대상**: 문자열 검증, 데이터 검증, 파일 검증 함수
+- **[상세 가이드](modules/validators.md)**
+
+### Security 모듈
+
+- **난이도**: 쉬움-중간
+- **특징**: 암호화/해시 함수, 대부분 순수 함수, 일부 외부 라이브러리 의존
+- **주요 테스트 대상**: 암호화/복호화, 해싱, 키 생성 기능
+- **[상세 가이드](modules/security.md)**
+
+### Config 모듈
+
+- **난이도**: 쉬움-중간
+- **특징**: 환경 변수 의존, 설정 로드 및 검증
+- **주요 테스트 대상**: 설정 로드, 기본값 적용, 환경 변수 오버라이드
+- **[상세 가이드](modules/config.md)**
+
+### Schemas 모듈
+
+- **난이도**: 쉬움-중간
+- **특징**: Pydantic 모델 기반, 데이터 검증 로직
+- **주요 테스트 대상**: 스키마 검증, 변환, 직렬화/역직렬화
+- **[상세 가이드](modules/schemas.md)**
+
+### Utils 모듈
+
+- **난이도**: 중간
+- **특징**: 다양한 유틸리티 함수, 일부 외부 의존성, 작은 함수들이 많음
+- **주요 테스트 대상**: 날짜 처리, 문자열 처리, 파일 처리 유틸리티, 페이지네이션
+- **테스트 방법**: 단위 테스트, 모의 객체, 경계값 테스트
+- **[상세 가이드](modules/utils.md)**
+
+### Exceptions 모듈
+
+- **난이도**: 중간
+- **특징**: 예외 클래스, 예외 처리기
+- **주요 테스트 대상**: 커스텀 예외, 예외 핸들러, 오류 응답
+- **[상세 가이드](modules/exceptions.md)**
+
+### Middleware 모듈
+
+- **난이도**: 중간-어려움
+- **특징**: 요청/응답 처리, HTTP 컨텍스트 의존성
+- **주요 테스트 대상**: 미들웨어 함수, 요청/응답 수정, 오류 처리
+- **[상세 가이드](modules/middleware.md)**
+
+### Auth 모듈
+
+- **난이도**: 중간-어려움
+- **특징**: 인증 흐름, 토큰 처리, 사용자 검증
+- **주요 테스트 대상**: 인증 의존성, 토큰 생성/검증, 권한 확인
+- **[상세 가이드](modules/auth.md)**
+
+### Cache 모듈
+
+- **난이도**: 어려움
+- **특징**: Redis 의존성, 비동기 코드, 캐싱 로직
+- **주요 테스트 대상**: 캐시 데코레이터, 캐시 무효화, 캐시 키 관리
+- **[상세 가이드](modules/cache.md)**
+
+### Database 모듈
+
+- **난이도**: 어려움
+- **특징**: DB 의존성, 트랜잭션, 비동기 ORM
+- **주요 테스트 대상**: 모델, 세션 관리, CRUD 작업
+- **[상세 가이드](modules/database.md)**
+
+### Monitoring 모듈
+
+- **난이도**: 어려움
+- **특징**: 시스템 상태 모니터링, 로깅, 외부 도구 의존성
+- **주요 테스트 대상**: 로깅 기능, 헬스 체크, 메트릭 수집
+- **[상세 가이드](modules/monitoring.md)**
+
+### API 모듈
+
+- **난이도**: 가장 어려움
+- **특징**: 전체 애플리케이션 스택, 여러 의존성 결합
+- **주요 테스트 대상**: HTTP 엔드포인트, 요청/응답 사이클, 라우터
+- **[상세 가이드](modules/api.md)**
+
+### Dependency 모듈
+
+- **난이도**: 중간-어려움
+- **특징**: 의존성 주입 시스템, 인증/권한 검증, 상태 관리
+- **주요 테스트 대상**: 의존성 함수, 의존성 오버라이드, 의존성 체인
+- **[상세 가이드](modules/dependency.md)**
+
+### Repositories 모듈
+
+- **난이도**: 중간
+- **특징**: 데이터베이스 액세스 계층, 쿼리 관리, 트랜잭션 처리
+- **주요 테스트 대상**: CRUD 연산, 복잡한 쿼리, 동시성 처리
+- **[상세 가이드](modules/repositories.md)**
+
+### Services 모듈
+
+- **난이도**: 중간
+- **특징**: 비즈니스 로직, 여러 리포지토리 조합, 외부 API 호출
+- **주요 테스트 대상**: 비즈니스 규칙, 데이터 변환, 트랜잭션 관리
+- **[상세 가이드](modules/services.md)**
+
+## 참고하기
+
+테스트를 작성할 때 다음 문서를 함께 참고하세요:
+
+- [테스트 개요](overview.md)
+- [테스트 실행 방법](running_tests.md)
+- [테스트 구조](structure.md)
+- [테스트 작성 모범 사례](best_practices.md)
+- [테스트 도구 가이드](tools.md)
+- [공통 테스트 패턴](common_test_patterns.md)
