+++ fastapi_template/docs/testing/overview.md
@@ -1,0 +1,47 @@
+# 테스트 개요
+
+이 문서는 FastAPI 템플릿 프로젝트의 테스트 방법론과 접근 방식에 대한 개요를 제공합니다.
+
+## 테스트 철학
+
+이 프로젝트는 다음과 같은 테스트 철학을 따릅니다:
+
+1. **테스트 자동화**: 모든 테스트는 자동화되어 CI/CD 파이프라인에서 실행될 수 있어야 합니다.
+2. **테스트 피라미드**: 단위 테스트 > 통합 테스트 > E2E 테스트 비율을 유지합니다.
+3. **격리된 테스트**: 각 테스트는 독립적으로 실행될 수 있어야 합니다.
+4. **테스트 가독성**: 테스트 코드는 명확하고 이해하기 쉬워야 합니다.
+
+## 테스트 유형
+
+이 프로젝트는 다음과 같은 테스트 유형을 지원합니다:
+
+1. **단위 테스트**: 개별 함수나 메서드의 기능 테스트
+2. **통합 테스트**: 여러 모듈 간의 상호작용 테스트
+3. **API 테스트**: HTTP 엔드포인트 테스트
+4. **E2E 테스트**: 전체 애플리케이션 흐름 테스트
+
+## 테스트 도구
+
+프로젝트에서 사용하는 주요 테스트 도구:
+
+- **pytest**: 기본 테스트 프레임워크
+- **pytest-asyncio**: 비동기 코드 테스트
+- **pytest-cov**: 코드 커버리지 측정
+- **TestClient**: FastAPI의 내장 테스트 클라이언트
+
+## 테스트 환경
+
+테스트는 다음과 같은 환경에서 실행될 수 있습니다:
+
+1. **로컬 개발 환경**: 개발자의 로컬 머신에서 실행
+2. **CI 환경**: GitHub Actions 또는 다른 CI 도구에서 자동으로 실행
+3. **스테이징 환경**: 배포 전 최종 검증을 위한 환경
+
+## 테스트 데이터 관리
+
+테스트 데이터는 다음과 같은 방식으로 관리됩니다:
+
+1. **픽스처(Fixture)**: pytest 픽스처를 사용하여 테스트 데이터 제공
+2. **팩토리(Factory)**: 테스트 객체 생성을 위한 팩토리 함수
+3. **모킹(Mocking)**: 외부 의존성에 대한 모의 객체 사용
+4. **인메모리 데이터베이스**: 테스트용 SQLite 인메모리 데이터베이스 활용
