+++ fastapi_template/py_project_tree.txt
@@ -41,6 +41,9 @@ fastapi_template\app\common\utils: __init__.py
 fastapi_template\app\common\utils: datetime.py
 fastapi_template\app\common\utils: pagination.py
 fastapi_template\app\common\validators: __init__.py
+fastapi_template\app\common\validators: data_validators.py
+fastapi_template\app\common\validators: file_validators.py
+fastapi_template\app\common\validators: string_validators.py
 fastapi_template\app\db\models: __init__.py
 fastapi_template\app\db\models: item.py
 fastapi_template\app\db\models: user.py
@@ -53,13 +56,35 @@ fastapi_template\app\services: item_service.py
 fastapi_template\app\services: user_service.py
 fastapi_template\app\tests: test_users.py
 fastapi_template\docs: architecture.md
-fastapi_template\docs: common_modules.md
-fastapi_template\docs: database.md
+fastapi_template\docs: common__overview.md
+fastapi_template\docs: common_auth.md
+fastapi_template\docs: common_cache.md
+fastapi_template\docs: common_config.md
+fastapi_template\docs: common_database.md
+fastapi_template\docs: common_exceptions.md
+fastapi_template\docs: common_middleware.md
+fastapi_template\docs: common_monitoring.md
+fastapi_template\docs: common_schemas.md
+fastapi_template\docs: common_security.md
+fastapi_template\docs: common_utils.md
+fastapi_template\docs: common_validators.md
 fastapi_template\docs: development_plan.md
+fastapi_template\docs: testing_guide.md
 fastapi_template\docs\database: 01-db-setup.md
 fastapi_template\docs\database: 02-models.md
 fastapi_template\docs\database: 03-migrations.md
 fastapi_template\docs\database: 04-schemas.md
 fastapi_template\docs\database: 05-services.md
 fastapi_template\docs\database: 06-docker.md
+fastapi_template\docs\testing: auth.md
+fastapi_template\docs\testing: best_practices.md
+fastapi_template\docs\testing: config.md
+fastapi_template\docs\testing: module_guides.md
+fastapi_template\docs\testing: overview.md
+fastapi_template\docs\testing: running_tests.md
+fastapi_template\docs\testing: security.md
+fastapi_template\docs\testing: structure.md
+fastapi_template\docs\testing: tools.md
+fastapi_template\docs\testing: utils.md
+fastapi_template\docs\testing: validators.md
 fastapi_template\tests: test_common_modules.py
